alphavantage/requirements.txt
python-dotenv
requests
pandas
polars
